text-summary-by-extraction/requirements.txt
requests==2.25.1
Django==3.1.6
snownlp==0.12.3
beautifulsoup4==4.9.3
mysql_connector_repackaged==0.3.1
